removed script dependence of tree generation
